Start pages refactoring

+++ Registrant/Forms/AddOrEditShipment.xaml.cs
@@ -238,7 +238,7 @@ namespace Registrant.Forms
                 {
                     using (RegistrantCoreContext ef = new RegistrantCoreContext())
                     {
-                        var shipment = ef.Shipments.FirstOrDefault(x => x.IdShipment == Convert.ToInt64(idcont.Text));
+                        var shipment = ef.Shipments.FirstOrDefault(x => x.IdShipment == Convert.ToInt32(idcont.Text));
 
                         if (shipment != null)
                         {
@@ -298,7 +298,7 @@ namespace Registrant.Forms
                 {
                     using (RegistrantCoreContext ef = new RegistrantCoreContext())
                     {
-                        var shipment = ef.Shipments.FirstOrDefault(x => x.IdShipment == Convert.ToInt64(idcont.Text));
+                        var shipment = ef.Shipments.FirstOrDefault(x => x.IdShipment == Convert.ToInt32(idcont.Text));
 
                         if (shipment != null)
                         {
